Move Footer component to footer subdirectory
Relocated Footer.tsx to src/components/layout/footer/Footer.tsx and updated import in Layout.tsx. Added an empty Footer.module.css file for future styling.
