+++ src/components/layout/Layout.tsx
@@ -1,5 +1,5 @@
 import Header from './header/Header'
-import Footer from './Footer'
+import Footer from './footer/Footer'
 
 export default function Layout({ children }: { children: React.ReactNode }) {
 	return (
